+++ apps/chat/src/hooks/files/tests/useDialFileManager.spec.tsx
@@ -11,13 +11,11 @@ import type { ListFilesItemDto } from '@epam/chat-api-client';
 import { ListFilesItemDtoNodeTypeEnum } from '@epam/chat-api-client';
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { DialFileManagerI18nKeys } from '../../../constants/translation-keys';
 import * as filesApi from '../../../server-api/files.api';
 import {
   DialFileManagerActionProfile,
   DialFileManagerVariant,
 } from '../../../types/file-manager-variant';
-import * as fileNameUtils from '../../../utils/file-name';
 import { useDialFileManager } from '../useDialFileManager';
 
 vi.mock('../../../server-api/files.api');
@@ -1046,127 +1044,6 @@ describe('useDialFileManager', () => {
     await act(async () => resolvePromise());
   });
 
-  describe('integration: sanitize → conflict resolution → upload mode', () => {
-    it('sanitizes file names via onValidateUpload before upload', async () => {
-      const mockSanitize = vi.mocked(fileNameUtils.sanitizeFileName);
-      mockSanitize.mockImplementation((name) => name.replace(/[/:]/g, '_'));
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      const files = [
-        {
-          name: 'report:final.pdf',
-          fileContent: new File([], 'report:final.pdf'),
-        },
-        {
-          name: 'data/export.csv',
-          fileContent: new File([], 'data/export.csv'),
-        },
-      ];
-
-      await act(async () => {
-        await result.current.onValidateUpload(files, [], '/My files');
-      });
-
-      expect(files[0].name).toBe('report_final.pdf');
-      expect(files[1].name).toBe('data_export.csv');
-      expect(mockSanitize).toHaveBeenCalledTimes(2);
-    });
-
-    it('selects overwrite for a file whose sanitized name matches the cache, create-only otherwise', async () => {
-      const mockUploadFile = vi.mocked(filesApi.uploadFile);
-      mockUploadFile.mockResolvedValue({ url: `files/${BUCKET}/file.pdf` });
-
-      const mockSanitize = vi.mocked(fileNameUtils.sanitizeFileName);
-      mockSanitize.mockImplementation((name) => name);
-
-      const cachedFile: ListFilesItemDto = {
-        name: 'existing.pdf',
-        path: 'existing.pdf',
-        folderId: `${BUCKET}:`,
-        nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-        bucket: BUCKET,
-      };
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [cachedFile],
-        permissions: ['READ', 'WRITE'],
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => {
-        result.current.onUploadFiles(
-          [
-            {
-              name: 'existing.pdf',
-              fileContent: new File(['data'], 'existing.pdf'),
-            },
-            {
-              name: 'new-file.pdf',
-              fileContent: new File(['data'], 'new-file.pdf'),
-            },
-          ],
-          '/My files',
-        );
-      });
-
-      await waitFor(() => expect(mockUploadFile).toHaveBeenCalledTimes(2));
-
-      const calls = mockUploadFile.mock.calls;
-      const existingCall = calls.find((c) => c[1] === 'existing.pdf');
-      const newCall = calls.find((c) => c[1] === 'new-file.pdf');
-
-      expect(existingCall?.[3]).toEqual(
-        expect.objectContaining({ uploadMode: 'overwrite' }),
-      );
-      expect(newCall?.[3]).toEqual(
-        expect.objectContaining({ uploadMode: 'create-only' }),
-      );
-    });
-
-    it('onValidateUpload always returns valid:true regardless of name collisions', async () => {
-      const existingFile: ListFilesItemDto = {
-        name: 'report.pdf',
-        path: 'report.pdf',
-        folderId: `${BUCKET}:`,
-        nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-        bucket: BUCKET,
-      };
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [existingFile],
-        permissions: ['READ', 'WRITE'],
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      let validation:
-        | Awaited<ReturnType<typeof result.current.onValidateUpload>>
-        | undefined;
-      await act(async () => {
-        validation = await result.current.onValidateUpload(
-          [{ name: 'report.pdf', fileContent: new File([], 'report.pdf') }],
-          [existingFile as never],
-          '/My files',
-        );
-      });
-
-      expect(validation).toEqual({ valid: true });
-    });
-  });
-
   describe('tab-aware behavior', () => {
     const sharedRootItem: ListFilesItemDto = {
       name: 'team-docs',
@@ -1177,35 +1054,6 @@ describe('useDialFileManager', () => {
       permissions: ['READ', 'WRITE'],
       author: 'Owner User',
     };
-
-    it('switching tabs resets folderPath to root and clears the listing cache', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: 'reports/',
-        items: [],
-        permissions: ['READ', 'WRITE'],
-      });
-
-      const { result, rerender } = renderHook(
-        ({ tab }: { tab: DialFileManagerTabs }) =>
-          useDialFileManager({ bucket: BUCKET, activeTab: tab }),
-        { initialProps: { tab: DialFileManagerTabs.MyFiles } },
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onPathChange('/My files/reports/'));
-      await waitFor(() =>
-        expect(result.current.path).toBe('/My files/reports/'),
-      );
-
-      mockListPublicFiles.mockClear();
-
-      rerender({ tab: DialFileManagerTabs.Organization });
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      expect(result.current.path).toBe('/My files');
-      expect(mockListPublicFiles).toHaveBeenCalled();
-    });
 
     describe('uploadEnabled matrix', () => {
       it('is always false on Organization tab', async () => {
@@ -1689,1583 +1537,6 @@ describe('useDialFileManager', () => {
         );
       });
     });
-
-    describe('sharedWithMeIds', () => {
-      it('is populated with virtual DialFile paths matching root Shared listing items', async () => {
-        mockListSharedFiles.mockResolvedValue({
-          bucket: '',
-          path: '',
-          items: [sharedRootItem],
-        });
-
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.Shared,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-        expect(result.current.sharedWithMeIds).toEqual([
-          '/My files/team-docs/',
-        ]);
-      });
-
-      it('is undefined on MyFiles tab', async () => {
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.MyFiles,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-        expect(result.current.sharedWithMeIds).toBeUndefined();
-      });
-
-      it('is undefined on Organization tab', async () => {
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.Organization,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-        expect(result.current.sharedWithMeIds).toBeUndefined();
-      });
-
-      it('matches the actual root item virtual path exposed via items', async () => {
-        mockListSharedFiles.mockResolvedValue({
-          bucket: '',
-          path: '',
-          items: [sharedRootItem],
-        });
-
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.Shared,
-            rootLabel: 'Shared with me',
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-        const rootFolder = result.current.items[0];
-        const teamDocsNode = rootFolder.items?.find(
-          (i) => i.name === 'team-docs',
-        );
-        expect(teamDocsNode).toBeDefined();
-        expect(result.current.sharedWithMeIds).toEqual([teamDocsNode?.path]);
-      });
-    });
-
-    describe('sharedByMePaths', () => {
-      it('is populated with virtual DialFile paths matching listSharedByMe items', async () => {
-        mockListSharedByMe.mockResolvedValue({
-          bucket: BUCKET,
-          path: '',
-          items: [
-            {
-              name: 'a.pdf',
-              path: `files/${BUCKET}/a.pdf`,
-              folderId: `${BUCKET}:`,
-              nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-              bucket: BUCKET,
-            },
-            {
-              name: 'nested.pdf',
-              path: `files/${BUCKET}/reports/2024/nested.pdf`,
-              folderId: `${BUCKET}:`,
-              nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-              bucket: BUCKET,
-            },
-            {
-              name: 'shared-folder',
-              path: `files/${BUCKET}/shared-folder/`,
-              folderId: `${BUCKET}:`,
-              nodeType: ListFilesItemDtoNodeTypeEnum.Folder,
-              bucket: BUCKET,
-            },
-          ],
-        });
-
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.MyFiles,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-        await waitFor(() =>
-          expect(result.current.sharedByMePaths.size).toBe(3),
-        );
-
-        expect(result.current.sharedByMePaths).toEqual(
-          new Set([
-            '/My files/a.pdf',
-            '/My files/reports/2024/nested.pdf',
-            '/My files/shared-folder/',
-          ]),
-        );
-      });
-
-      it('is empty on the Shared tab', async () => {
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.Shared,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-        expect(result.current.sharedByMePaths.size).toBe(0);
-        expect(mockListSharedByMe).not.toHaveBeenCalled();
-      });
-
-      it('is empty on the Organization tab', async () => {
-        const { result } = renderHook(() =>
-          useDialFileManager({
-            bucket: BUCKET,
-            activeTab: DialFileManagerTabs.Organization,
-          }),
-        );
-        await waitFor(() => expect(result.current.isLoading).toBe(false));
-        expect(result.current.sharedByMePaths.size).toBe(0);
-      });
-    });
-  });
-
-  describe('onCreateFolderValidate', () => {
-    const renderAndWait = async (opts?: object) => {
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, ...opts }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      return result;
-    };
-
-    const parentFolder = {
-      id: 'root',
-      name: 'My files',
-      path: '/My files',
-      parentPath: '',
-      nodeType: DialFileNodeType.FOLDER,
-      folderId: BUCKET,
-      bucket: BUCKET,
-      items: [],
-    };
-
-    it('returns null for a valid folder name', async () => {
-      const result = await renderAndWait();
-      expect(
-        result.current.onCreateFolderValidate('reports', parentFolder),
-      ).toBeNull();
-    });
-
-    it('returns invalid chars error for a folder name containing colon', async () => {
-      const result = await renderAndWait({
-        forbiddenSymbolsRegExp: /[:]/,
-      });
-      const msg = result.current.onCreateFolderValidate(
-        'reports:2026',
-        parentFolder,
-      );
-      expect(msg).toBe(DialFileManagerI18nKeys.FolderNameInvalidChars);
-    });
-
-    it('returns invalid chars error for a folder name containing a path separator', async () => {
-      const result = await renderAndWait();
-      const msg = result.current.onCreateFolderValidate(
-        'reports/2026',
-        parentFolder,
-      );
-      expect(msg).toBe(DialFileManagerI18nKeys.FolderNameInvalidChars);
-    });
-  });
-
-  describe('onRenameValidate', () => {
-    const renderAndWait = async (opts?: object) => {
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, ...opts }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      return result;
-    };
-
-    const dummyItem = {
-      id: 'file.pdf',
-      name: 'file.pdf',
-      path: '/My files/file.pdf',
-      parentPath: '/My files',
-      nodeType: DialFileNodeType.ITEM,
-      folderId: BUCKET,
-      bucket: BUCKET,
-      items: [],
-    };
-
-    it('returns null for a valid name', async () => {
-      const result = await renderAndWait();
-      expect(
-        result.current.onRenameValidate('report.pdf', dummyItem),
-      ).toBeNull();
-    });
-
-    it('returns empty-name error for an empty name', async () => {
-      const result = await renderAndWait();
-      expect(result.current.onRenameValidate('', dummyItem)).toBeTruthy();
-    });
-
-    it('returns reserved name error for ".dial_folder"', async () => {
-      const result = await renderAndWait();
-      const msg = result.current.onRenameValidate(HIDDEN_FILE, dummyItem);
-      expect(msg).toBeTruthy();
-    });
-
-    it('returns invalid chars error for name containing forward slash', async () => {
-      const result = await renderAndWait();
-      const msg = result.current.onRenameValidate('a/b', dummyItem);
-      expect(msg).toBe(DialFileManagerI18nKeys.ForbiddenSymbolsTooltip);
-    });
-
-    it('returns invalid chars error for name containing backslash', async () => {
-      const result = await renderAndWait();
-      const msg = result.current.onRenameValidate('a\\b', dummyItem);
-      expect(msg).toBe(DialFileManagerI18nKeys.ForbiddenSymbolsTooltip);
-    });
-
-    it('returns forbidden-symbols error for name containing colon', async () => {
-      const result = await renderAndWait({
-        forbiddenSymbolsRegExp: /[:]/,
-      });
-      const msg = result.current.onRenameValidate('file:name.pdf', dummyItem);
-      expect(msg).toBe(DialFileManagerI18nKeys.ForbiddenSymbolsTooltip);
-    });
-
-    it('returns too-long error for name longer than 255 chars', async () => {
-      const result = await renderAndWait();
-      const msg = result.current.onRenameValidate('a'.repeat(256), dummyItem);
-      expect(msg).toBeTruthy();
-    });
-
-    it('returns forbidden-symbols error when name matches forbiddenSymbolsRegExp', async () => {
-      const result = await renderAndWait({
-        forbiddenSymbolsRegExp: /[<>]/,
-      });
-      const msg = result.current.onRenameValidate('file<name>', dummyItem);
-      expect(msg).toBe(DialFileManagerI18nKeys.ForbiddenSymbolsTooltip);
-    });
-
-    it('returns folder invalid chars error when renaming a folder with a forbidden symbol', async () => {
-      const result = await renderAndWait({
-        forbiddenSymbolsRegExp: /[:]/,
-      });
-      const folderItem = {
-        ...dummyItem,
-        name: 'reports',
-        path: '/My files/reports',
-        nodeType: DialFileNodeType.FOLDER,
-      };
-      const msg = result.current.onRenameValidate('reports:2026', folderItem);
-      expect(msg).toBe(DialFileManagerI18nKeys.FolderNameInvalidChars);
-    });
-
-    it('returns null when name does not match forbiddenSymbolsRegExp', async () => {
-      const result = await renderAndWait({
-        forbiddenSymbolsRegExp: /[<>]/,
-      });
-      expect(
-        result.current.onRenameValidate('valid.pdf', dummyItem),
-      ).toBeNull();
-    });
-
-    it('returns duplicate-name error for a sibling with the same name', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: 'report.pdf',
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-
-      const result = await renderAndWait();
-
-      expect(
-        result.current.onRenameValidate('REPORT.PDF', dummyItem),
-      ).toBeTruthy();
-    });
-  });
-
-  describe('onMoveToFiles', () => {
-    const mockRenameFiles = vi.mocked(filesApi.renameFiles);
-    const mockNotification = vi.fn();
-
-    beforeEach(() => {
-      mockRenameFiles.mockResolvedValue({ results: [] });
-    });
-
-    const renderAndWait = async () => {
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          onNotification: mockNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      return result;
-    };
-
-    it('calls renameFiles and triggers cache invalidation and retry on success', async () => {
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'file.pdf',
-            destinationPath: 'renamed.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/file.pdf',
-              destinationUrl: '/My files/renamed.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() => expect(mockRenameFiles).toHaveBeenCalledOnce());
-      expect(mockListFiles).toHaveBeenCalledTimes(2);
-    });
-
-    it('shows partial error toast when some items fail', async () => {
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          { sourcePath: 'a.pdf', destinationPath: 'a2.pdf', success: true },
-          {
-            sourcePath: 'b.pdf',
-            destinationPath: 'b2.pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/a.pdf',
-              destinationUrl: '/My files/a2.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/b.pdf',
-              destinationUrl: '/My files/b2.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('shows error toast when all items fail', async () => {
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'file.pdf',
-            destinationPath: 'renamed.pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/file.pdf',
-              destinationUrl: '/My files/renamed.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('navigates to the destination path after successfully renaming the current folder', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'reports',
-            path: 'reports/',
-            folderId: `${BUCKET}:reports/`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Folder,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'reports/',
-            destinationPath: 'archive/',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-      act(() => result.current.onPathChange('/My files/reports/'));
-      await waitFor(() =>
-        expect(result.current.path).toBe('/My files/reports/'),
-      );
-
-      act(() => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/',
-              destinationUrl: '/My files/archive/',
-              nodeType: DialFileNodeType.FOLDER,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() =>
-        expect(result.current.path).toBe('/My files/archive/'),
-      );
-    });
-
-    it('keeps the current path when a current-folder rename fails', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'reports',
-            path: 'reports/',
-            folderId: `${BUCKET}:reports/`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Folder,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'reports/',
-            destinationPath: 'archive/',
-            success: false,
-            error: 'Conflict',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-      act(() => result.current.onPathChange('/My files/reports/'));
-      await waitFor(() =>
-        expect(result.current.path).toBe('/My files/reports/'),
-      );
-
-      act(() => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/',
-              destinationUrl: '/My files/archive/',
-              nodeType: DialFileNodeType.FOLDER,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() => expect(mockRenameFiles).toHaveBeenCalledOnce());
-      expect(result.current.path).toBe('/My files/reports/');
-    });
-
-    it('calls only moveFiles for a cross-folder batch, not renameFiles', async () => {
-      mockMoveFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'inbox/draft.pdf',
-            destinationPath: 'reports/draft.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/inbox/draft.pdf',
-              destinationUrl: '/My files/reports/draft.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/inbox',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() => expect(mockMoveFiles).toHaveBeenCalledOnce());
-      expect(mockRenameFiles).not.toHaveBeenCalled();
-      expect(mockNotification).toHaveBeenCalledWith({
-        variant: NotificationVariant.Success,
-        title: 'dialFileManager.itemMovedSuccessfully',
-        message: 'dialFileManager.itemMovedToFolder',
-      });
-    });
-
-    it('shows a success toast with the moved item count for multiple files', async () => {
-      mockMoveFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'inbox/a.pdf',
-            destinationPath: 'reports/a.pdf',
-            success: true,
-          },
-          {
-            sourcePath: 'inbox/b.pdf',
-            destinationPath: 'reports/b.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/inbox/a.pdf',
-              destinationUrl: '/My files/reports/a.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/inbox/b.pdf',
-              destinationUrl: '/My files/reports/b.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/inbox',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith({
-          variant: NotificationVariant.Success,
-          title: 'dialFileManager.itemsMovedSuccessfully',
-          message: 'dialFileManager.itemsMovedToFolder',
-        }),
-      );
-    });
-
-    it('passes overwrite=true from conflict resolution to moveFiles', async () => {
-      mockMoveFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'inbox/draft.pdf',
-            destinationPath: 'reports/draft.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/inbox/draft.pdf',
-              destinationUrl: '/My files/reports/draft.pdf',
-              overwrite: true,
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/inbox',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() => expect(mockMoveFiles).toHaveBeenCalledOnce());
-      expect(mockMoveFiles).toHaveBeenCalledWith(
-        [
-          expect.objectContaining({
-            destinationPath: 'reports/draft.pdf',
-            overwrite: true,
-          }),
-        ],
-        expect.anything(),
-      );
-    });
-
-    it('calls both renameFiles and moveFiles for a mixed batch and merges the failure toast', async () => {
-      mockRenameFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'a.pdf',
-            destinationPath: 'a2.pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-      mockMoveFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'inbox/draft.pdf',
-            destinationPath: 'reports/draft.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onMoveToFiles(
-          [
-            {
-              sourceUrl: '/My files/a.pdf',
-              destinationUrl: '/My files/a2.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/inbox/draft.pdf',
-              destinationUrl: '/My files/reports/draft.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-          '/My files',
-        );
-      });
-
-      await waitFor(() => {
-        expect(mockRenameFiles).toHaveBeenCalledOnce();
-        expect(mockMoveFiles).toHaveBeenCalledOnce();
-      });
-      expect(mockNotification).toHaveBeenCalledWith(
-        expect.objectContaining({ variant: NotificationVariant.Error }),
-      );
-    });
-  });
-
-  describe('onCopyFiles', () => {
-    const mockNotification = vi.fn();
-
-    const renderAndWait = async () => {
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          onNotification: mockNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      return result;
-    };
-
-    it('invalidates cache and shows a success toast on full copy success', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'reports/q1.pdf',
-            destinationPath: 'archive/q1.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/q1.pdf',
-              destinationUrl: '/My files/archive/q1.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/archive',
-        );
-      });
-
-      await waitFor(() => expect(mockCopyFiles).toHaveBeenCalledOnce());
-      expect(mockListFiles).toHaveBeenCalledTimes(2);
-      expect(mockNotification).toHaveBeenCalledWith({
-        variant: NotificationVariant.Success,
-        title: 'dialFileManager.itemCopiedSuccessfully',
-        message: 'dialFileManager.itemCopiedToFolder',
-      });
-    });
-
-    it('shows a success toast with the copied item count for multiple files', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'a.pdf',
-            destinationPath: 'archive/a.pdf',
-            success: true,
-          },
-          {
-            sourcePath: 'b.pdf',
-            destinationPath: 'archive/b.pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/a.pdf',
-              destinationUrl: '/My files/archive/a.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/b.pdf',
-              destinationUrl: '/My files/archive/b.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/archive',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith({
-          variant: NotificationVariant.Success,
-          title: 'dialFileManager.itemsCopiedSuccessfully',
-          message: 'dialFileManager.itemsCopiedToFolder',
-        }),
-      );
-    });
-
-    it('collapses a double-slash destinationUrl (folder prefix + leading slash) before sending', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'Folder_for_test_copy/img.png',
-            destinationPath: 'folder_for_test_copy_1/img.png',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/Folder_for_test_copy/img.png',
-              destinationUrl: '/My files/folder_for_test_copy_1//img.png',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/folder_for_test_copy_1',
-        );
-      });
-
-      await waitFor(() => expect(mockCopyFiles).toHaveBeenCalledOnce());
-      expect(mockCopyFiles).toHaveBeenCalledWith(
-        [
-          expect.objectContaining({
-            destinationPath: 'folder_for_test_copy_1/img.png',
-          }),
-        ],
-        expect.anything(),
-      );
-    });
-
-    it('passes overwrite=true from conflict resolution to copyFiles', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'requirements.txt',
-            destinationPath: 'Folder1/requirements.txt',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/requirements.txt',
-              destinationUrl: '/My files/Folder1/requirements.txt',
-              overwrite: true,
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/Folder1',
-        );
-      });
-
-      await waitFor(() => expect(mockCopyFiles).toHaveBeenCalledOnce());
-      expect(mockCopyFiles).toHaveBeenCalledWith(
-        [
-          expect.objectContaining({
-            destinationPath: 'Folder1/requirements.txt',
-            overwrite: true,
-          }),
-        ],
-        expect.anything(),
-      );
-    });
-
-    it('shows a partial-failure toast with the failed count', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          { sourcePath: 'a.pdf', destinationPath: 'a2.pdf', success: true },
-          {
-            sourcePath: 'b.pdf',
-            destinationPath: 'b2.pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/a.pdf',
-              destinationUrl: '/My files/a2.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/b.pdf',
-              destinationUrl: '/My files/b2.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('shows a full-failure toast when every item fails', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'reports/q1.pdf',
-            destinationPath: 'archive/q1.pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/q1.pdf',
-              destinationUrl: '/My files/archive/q1.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/archive',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('handles a same-folder destination (duplicate) correctly on success', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'reports/q1.pdf',
-            destinationPath: 'reports/q1 (1).pdf',
-            success: true,
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/q1.pdf',
-              destinationUrl: '/My files/reports/q1 (1).pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() => expect(mockCopyFiles).toHaveBeenCalledOnce());
-      expect(mockCopyFiles).toHaveBeenCalledWith(
-        [
-          expect.objectContaining({
-            bucket: BUCKET,
-            sourcePath: 'reports/q1.pdf',
-            destinationPath: 'reports/q1 (1).pdf',
-          }),
-        ],
-        expect.anything(),
-      );
-      // Source and destination share the same parent folder — invalidated once.
-      expect(mockListFiles).toHaveBeenCalledTimes(2);
-      expect(mockNotification).toHaveBeenCalledWith({
-        variant: NotificationVariant.Success,
-        title: 'dialFileManager.itemCopiedSuccessfully',
-        message: 'dialFileManager.itemCopiedToFolder',
-      });
-    });
-
-    it('shows the existing partial-failure toast for a same-folder destination (duplicate)', async () => {
-      mockCopyFiles.mockResolvedValue({
-        results: [
-          {
-            sourcePath: 'a.pdf',
-            destinationPath: 'a (1).pdf',
-            success: true,
-          },
-          {
-            sourcePath: 'b.pdf',
-            destinationPath: 'b (1).pdf',
-            success: false,
-            error: 'Forbidden',
-          },
-        ],
-      });
-
-      const result = await renderAndWait();
-
-      await act(async () => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/a.pdf',
-              destinationUrl: '/My files/a (1).pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-            {
-              sourceUrl: '/My files/b.pdf',
-              destinationUrl: '/My files/b (1).pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockNotification).toHaveBeenCalledWith(
-          expect.objectContaining({
-            variant: NotificationVariant.Error,
-            message: 'dialFileManager.copyPartialError',
-          }),
-        ),
-      );
-    });
-
-    it('clears isCopying with no toast when cancelled', async () => {
-      mockCopyFiles.mockImplementation(
-        (_items, signal) =>
-          new Promise((_resolve, reject) => {
-            signal?.addEventListener('abort', () => {
-              reject(new DOMException('Aborted', 'AbortError'));
-            });
-          }),
-      );
-
-      const result = await renderAndWait();
-
-      act(() => {
-        result.current.onCopyFiles(
-          [
-            {
-              sourceUrl: '/My files/reports/q1.pdf',
-              destinationUrl: '/My files/archive/q1.pdf',
-              nodeType: DialFileNodeType.ITEM,
-            },
-          ],
-          '/My files/archive',
-        );
-      });
-
-      await waitFor(() => expect(result.current.isCopying).toBe(true));
-
-      await act(async () => {
-        result.current.cancelCopyMove();
-      });
-
-      await waitFor(() => expect(result.current.isCopying).toBe(false));
-      expect(mockNotification).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('onSearchFiles', () => {
-    it('debounces search and calls listFiles only once after 300ms', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: 'report.pdf',
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      const initialCallCount = mockListFiles.mock.calls.length;
-
-      vi.useFakeTimers();
-      try {
-        act(() => result.current.onSearchFiles('/', 'rep'));
-        act(() => result.current.onSearchFiles('/', 'repo'));
-        act(() => result.current.onSearchFiles('/', 'repor'));
-
-        expect(mockListFiles).toHaveBeenCalledTimes(initialCallCount);
-
-        act(() => {
-          vi.advanceTimersByTime(300);
-        });
-        await act(() => Promise.resolve());
-
-        expect(mockListFiles).toHaveBeenCalledTimes(initialCallCount + 1);
-        const lastCall =
-          mockListFiles.mock.calls[mockListFiles.mock.calls.length - 1][0];
-        expect(lastCall).toMatchObject({ recursive: true });
-        expect(result.current.searchResults).toHaveLength(1);
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it('cancels in-flight search when a second query arrives before the first resolves', async () => {
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      const initialCallCount = mockListFiles.mock.calls.length;
-
-      let resolveFirst!: (value: {
-        bucket: string;
-        path: string;
-        items: ListFilesItemDto[];
-        nextToken: undefined;
-      }) => void;
-      const firstPromise = new Promise<{
-        bucket: string;
-        path: string;
-        items: ListFilesItemDto[];
-        nextToken: undefined;
-      }>((res) => {
-        resolveFirst = res;
-      });
-      mockListFiles.mockReturnValueOnce(firstPromise).mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [],
-        nextToken: undefined,
-      });
-
-      vi.useFakeTimers();
-      try {
-        act(() => result.current.onSearchFiles('/', 'first'));
-        act(() => {
-          vi.advanceTimersByTime(300);
-        });
-
-        act(() => result.current.onSearchFiles('/', 'second'));
-        act(() => {
-          vi.advanceTimersByTime(300);
-        });
-
-        // Flush second fetch (empty result)
-        await act(() => Promise.resolve());
-
-        expect(mockListFiles).toHaveBeenCalledTimes(initialCallCount + 2);
-        expect(result.current.isSearching).toBe(false);
-
-        // Resolve the stale first fetch — its result should be ignored
-        resolveFirst({
-          bucket: BUCKET,
-          path: '',
-          items: [
-            {
-              name: 'first-only.pdf',
-              path: 'first-only.pdf',
-              folderId: `${BUCKET}:`,
-              nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-              bucket: BUCKET,
-            },
-          ],
-          nextToken: undefined,
-        });
-        await act(() => Promise.resolve());
-
-        expect(
-          result.current.searchResults?.some(
-            (r) => r.name === 'first-only.pdf',
-          ),
-        ).toBe(false);
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-
-    it('reconstructs virtual path for a nested file returned by search', async () => {
-      const nestedItem: ListFilesItemDto = {
-        name: 'summary.pdf',
-        path: 'reports/q1/summary.pdf',
-        url: `files/${BUCKET}/reports/q1/summary.pdf`,
-        folderId: `${BUCKET}:reports/q1/`,
-        nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-        bucket: BUCKET,
-      };
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [nestedItem],
-        nextToken: undefined,
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      vi.useFakeTimers();
-      try {
-        act(() => result.current.onSearchFiles('/', 'summary'));
-        act(() => {
-          vi.advanceTimersByTime(300);
-        });
-        await act(() => Promise.resolve());
-
-        const found = result.current.searchResults?.find(
-          (r) => r.name === 'summary.pdf',
-        );
-        expect(found).toBeDefined();
-        expect(found?.path).toBe('/My files/reports/q1/summary.pdf');
-        expect(found?.parentPath).toBe('/My files/reports/q1');
-      } finally {
-        vi.useRealTimers();
-      }
-    });
-  });
-
-  describe('onManagePermissions', () => {
-    it('resolves the correct item and opens the share modal target', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: `files/${BUCKET}/report.pdf`,
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      expect(result.current.shareTarget).toBeNull();
-
-      act(() =>
-        result.current.onManagePermissions(`files/${BUCKET}/report.pdf`),
-      );
-
-      expect(result.current.shareTarget).toEqual({
-        bucket: BUCKET,
-        path: 'report.pdf',
-        name: 'report.pdf',
-      });
-    });
-
-    it('does nothing when path does not resolve to a loaded item', async () => {
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onManagePermissions('unknown-path'));
-
-      expect(result.current.shareTarget).toBeNull();
-    });
-
-    it('onCloseShareModal clears the share target', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: `files/${BUCKET}/report.pdf`,
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() =>
-        result.current.onManagePermissions(`files/${BUCKET}/report.pdf`),
-      );
-      expect(result.current.shareTarget).not.toBeNull();
-
-      act(() => result.current.onCloseShareModal());
-      expect(result.current.shareTarget).toBeNull();
-    });
-
-    it('onCreateShareLink calls shareFiles with the resolved target and permission', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: `files/${BUCKET}/report.pdf`,
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-      mockShareFiles.mockResolvedValue({
-        invitationLink: 'https://chat.example.com/share/abc',
-      });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() =>
-        result.current.onManagePermissions(`files/${BUCKET}/report.pdf`),
-      );
-
-      let link: string | undefined;
-      await act(async () => {
-        link = await result.current.onCreateShareLink('read');
-      });
-
-      expect(mockShareFiles).toHaveBeenCalledWith(
-        [{ bucket: BUCKET, path: 'report.pdf' }],
-        'read',
-        expect.any(AbortSignal),
-      );
-      expect(link).toBe('https://chat.example.com/share/abc');
-    });
-
-    it('clears isSharing immediately when the modal is closed on a pending request', async () => {
-      mockListFiles.mockResolvedValue({
-        bucket: BUCKET,
-        path: '',
-        items: [
-          {
-            name: 'report.pdf',
-            path: `files/${BUCKET}/report.pdf`,
-            folderId: `${BUCKET}:`,
-            nodeType: ListFilesItemDtoNodeTypeEnum.Item,
-            bucket: BUCKET,
-          },
-        ],
-        nextToken: undefined,
-      });
-      mockShareFiles.mockImplementation(
-        (_items, _permission, signal) =>
-          new Promise((_resolve, reject) => {
-            signal?.addEventListener('abort', () => {
-              reject(new DOMException('Aborted', 'AbortError'));
-            });
-          }),
-      );
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() =>
-        result.current.onManagePermissions(`files/${BUCKET}/report.pdf`),
-      );
-
-      let pendingRejection: Promise<string> | undefined;
-      act(() => {
-        pendingRejection = result.current.onCreateShareLink('read');
-        pendingRejection.catch(() => undefined);
-      });
-
-      await waitFor(() => expect(result.current.isSharing).toBe(true));
-
-      act(() => result.current.onCloseShareModal());
-
-      expect(result.current.isSharing).toBe(false);
-      expect(result.current.shareTarget).toBeNull();
-      await expect(pendingRejection).rejects.toThrow('Aborted');
-    });
-  });
-
-  describe('onUnshareFiles and onRemoveFilesAccess', () => {
-    const sharedWithMeFile = {
-      id: `files/${OWNER_BUCKET}/team-docs/`,
-      name: 'team-docs',
-      path: '/Shared with me/team-docs/',
-      parentPath: '/Shared with me',
-      nodeType: DialFileNodeType.FOLDER,
-      folderId: `${OWNER_BUCKET}:files/${OWNER_BUCKET}/team-docs/`,
-      bucket: OWNER_BUCKET,
-    };
-
-    const myOwnedFile = {
-      id: `files/${BUCKET}/report.pdf`,
-      name: 'report.pdf',
-      path: '/My files/report.pdf',
-      parentPath: '/My files',
-      nodeType: DialFileNodeType.ITEM,
-      folderId: `${BUCKET}:`,
-      bucket: BUCKET,
-    };
-
-    it('onUnshareFiles calls discardShared and triggers a retry with no toast on success', async () => {
-      mockDiscardShared.mockResolvedValue({ success: true });
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Shared,
-          actionProfile: DialFileManagerActionProfile.Full,
-          onNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      mockListSharedFiles.mockClear();
-
-      act(() => result.current.onUnshareFiles([sharedWithMeFile]));
-
-      await waitFor(() =>
-        expect(mockDiscardShared).toHaveBeenCalledWith([
-          { bucket: OWNER_BUCKET, path: 'team-docs/' },
-        ]),
-      );
-      await waitFor(() => expect(mockListSharedFiles).toHaveBeenCalled());
-      expect(onNotification).not.toHaveBeenCalled();
-    });
-
-    it('onUnshareFiles shows an error toast on failure', async () => {
-      mockDiscardShared.mockRejectedValue(new Error('failed'));
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Shared,
-          actionProfile: DialFileManagerActionProfile.Full,
-          onNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onUnshareFiles([sharedWithMeFile]));
-
-      await waitFor(() =>
-        expect(onNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('onRemoveFilesAccess calls revokeAccess and triggers a retry with no toast on success', async () => {
-      mockRevokeAccess.mockResolvedValue({ success: true });
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.MyFiles,
-          actionProfile: DialFileManagerActionProfile.Full,
-          onNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-      mockListFiles.mockClear();
-
-      act(() => result.current.onRemoveFilesAccess([myOwnedFile]));
-
-      await waitFor(() =>
-        expect(mockRevokeAccess).toHaveBeenCalledWith([
-          { bucket: BUCKET, path: 'report.pdf' },
-        ]),
-      );
-      await waitFor(() => expect(mockListFiles).toHaveBeenCalled());
-      expect(onNotification).not.toHaveBeenCalled();
-    });
-
-    it('onRemoveFilesAccess shows an error toast on failure', async () => {
-      mockRevokeAccess.mockRejectedValue(new Error('failed'));
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.MyFiles,
-          actionProfile: DialFileManagerActionProfile.Full,
-          onNotification,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onRemoveFilesAccess([myOwnedFile]));
-
-      await waitFor(() =>
-        expect(onNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-    });
-
-    it('resolves bucket/path correctly for a batch of items', async () => {
-      mockDiscardShared.mockResolvedValue({ success: true });
-
-      const secondSharedFile = {
-        id: `files/${OWNER_BUCKET}/notes.txt`,
-        name: 'notes.txt',
-        path: '/Shared with me/notes.txt',
-        parentPath: '/Shared with me',
-        nodeType: DialFileNodeType.ITEM,
-        folderId: `${OWNER_BUCKET}:`,
-        bucket: OWNER_BUCKET,
-      };
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Shared,
-          actionProfile: DialFileManagerActionProfile.Full,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() =>
-        result.current.onUnshareFiles([sharedWithMeFile, secondSharedFile]),
-      );
-
-      await waitFor(() =>
-        expect(mockDiscardShared).toHaveBeenCalledWith([
-          { bucket: OWNER_BUCKET, path: 'team-docs/' },
-          { bucket: OWNER_BUCKET, path: 'notes.txt' },
-        ]),
-      );
-    });
   });
 
   describe('isAnyOperationInProgress', () => {
@@ -3715,303 +1986,6 @@ describe('useDialFileManager', () => {
 
       await waitFor(() => expect(result.current.isMoving).toBe(false));
       expect(result.current.isAnyOperationInProgress).toBe(false);
-    });
-  });
-
-  describe('onGetInfo and clearMetadata', () => {
-    const myFilesItem = {
-      id: `files/${BUCKET}/report.pdf`,
-      name: 'report.pdf',
-      path: '/My files/report.pdf',
-      parentPath: '/My files',
-      nodeType: DialFileNodeType.ITEM,
-      folderId: `${BUCKET}:`,
-      bucket: BUCKET,
-    };
-
-    const metadataResponse = {
-      name: 'report.pdf',
-      nodeType: 'item',
-      bucket: BUCKET,
-      contentLength: 1234,
-      contentType: 'application/pdf',
-      author: 'Jane Doe',
-      permissions: ['READ', 'WRITE'],
-      updatedAt: 1700000000000,
-    };
-
-    it('resolves the current user bucket for a my_files item', async () => {
-      mockGetFileMetadata.mockResolvedValue(metadataResponse);
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.MyFiles,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(myFilesItem));
-      expect(result.current.isFileMetadataLoading).toBe(true);
-
-      await waitFor(() =>
-        expect(mockGetFileMetadata).toHaveBeenCalledWith({
-          bucket: BUCKET,
-          path: 'report.pdf',
-        }),
-      );
-      await waitFor(() =>
-        expect(result.current.isFileMetadataLoading).toBe(false),
-      );
-      expect(result.current.fileMetadata).toMatchObject({
-        path: '/My files/report.pdf',
-        contentLength: 1234,
-        author: 'Jane Doe',
-      });
-    });
-
-    it('resolves the owner bucket for a root-level shared item', async () => {
-      mockGetFileMetadata.mockResolvedValue(metadataResponse);
-      const sharedRootFile = {
-        id: `files/${OWNER_BUCKET}/notes.txt`,
-        name: 'notes.txt',
-        path: '/Shared with me/notes.txt',
-        parentPath: '/Shared with me',
-        nodeType: DialFileNodeType.ITEM,
-        folderId: `${OWNER_BUCKET}:`,
-        bucket: OWNER_BUCKET,
-      };
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Shared,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(sharedRootFile));
-
-      await waitFor(() =>
-        expect(mockGetFileMetadata).toHaveBeenCalledWith({
-          bucket: OWNER_BUCKET,
-          path: 'notes.txt',
-        }),
-      );
-    });
-
-    it('resolves the owner bucket for a nested shared item', async () => {
-      mockGetFileMetadata.mockResolvedValue(metadataResponse);
-      const nestedSharedFile = {
-        id: `files/${OWNER_BUCKET}/team-docs/report.pdf`,
-        name: 'report.pdf',
-        path: '/Shared with me/team-docs/report.pdf',
-        parentPath: '/Shared with me/team-docs',
-        nodeType: DialFileNodeType.ITEM,
-        folderId: `${OWNER_BUCKET}:files/${OWNER_BUCKET}/team-docs/`,
-        bucket: OWNER_BUCKET,
-      };
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Shared,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(nestedSharedFile));
-
-      await waitFor(() =>
-        expect(mockGetFileMetadata).toHaveBeenCalledWith({
-          bucket: OWNER_BUCKET,
-          path: 'team-docs/report.pdf',
-        }),
-      );
-    });
-
-    it('resolves the item bucket for an organization item', async () => {
-      mockGetFileMetadata.mockResolvedValue(metadataResponse);
-      const publicBucket = 'public-bucket';
-      const orgFile = {
-        id: `files/${publicBucket}/guide.pdf`,
-        name: 'guide.pdf',
-        path: '/Organization/guide.pdf',
-        parentPath: '/Organization',
-        nodeType: DialFileNodeType.ITEM,
-        folderId: `${publicBucket}:`,
-        bucket: publicBucket,
-      };
-
-      const { result } = renderHook(() =>
-        useDialFileManager({
-          bucket: BUCKET,
-          activeTab: DialFileManagerTabs.Organization,
-        }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(orgFile));
-
-      await waitFor(() =>
-        expect(mockGetFileMetadata).toHaveBeenCalledWith({
-          bucket: publicBucket,
-          path: 'guide.pdf',
-        }),
-      );
-    });
-
-    it('shows an error toast and clears loading when getFileMetadata rejects', async () => {
-      mockGetFileMetadata.mockRejectedValue(new Error('failed'));
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, onNotification }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(myFilesItem));
-
-      await waitFor(() =>
-        expect(onNotification).toHaveBeenCalledWith(
-          expect.objectContaining({ variant: NotificationVariant.Error }),
-        ),
-      );
-      expect(result.current.isFileMetadataLoading).toBe(false);
-    });
-
-    it('clearMetadata resets fileMetadata and isFileMetadataLoading', async () => {
-      mockGetFileMetadata.mockResolvedValue(metadataResponse);
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => result.current.onGetInfo(myFilesItem));
-      await waitFor(() =>
-        expect(result.current.fileMetadata).not.toBeUndefined(),
-      );
-
-      act(() => result.current.clearMetadata());
-
-      expect(result.current.fileMetadata).toBeUndefined();
-      expect(result.current.isFileMetadataLoading).toBe(false);
-    });
-  });
-
-  describe('onUploadArchive', () => {
-    const mockUploadArchive = vi.mocked(filesApi.uploadArchive);
-
-    it('invalidates the destination cache and shows no toast on full success', async () => {
-      mockUploadArchive.mockResolvedValue({
-        results: [
-          { path: 'reports/a.txt', success: true },
-          { path: 'reports/b.txt', success: true },
-        ],
-      });
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, onNotification }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => {
-        result.current.onUploadArchive(
-          new File(['zip'], 'archive.zip'),
-          'archive.zip',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() => expect(result.current.uploadBatchState).toBeNull());
-      expect(onNotification).not.toHaveBeenCalled();
-    });
-
-    it('shows a partial-failure toast with the failed count', async () => {
-      mockUploadArchive.mockResolvedValue({
-        results: [
-          { path: 'reports/a.txt', success: true },
-          { path: 'reports/b.txt', success: false, error: 'Conflict' },
-        ],
-      });
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, onNotification }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => {
-        result.current.onUploadArchive(
-          new File(['zip'], 'archive.zip'),
-          'archive.zip',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() =>
-        expect(onNotification).toHaveBeenCalledWith(
-          expect.objectContaining({
-            variant: NotificationVariant.Error,
-            message: 'dialFileManager.uploadArchivePartialError',
-          }),
-        ),
-      );
-    });
-
-    it('shows a full-failure toast when the request rejects', async () => {
-      mockUploadArchive.mockRejectedValue(new Error('network error'));
-      const onNotification = vi.fn();
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET, onNotification }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      act(() => {
-        result.current.onUploadArchive(
-          new File(['zip'], 'archive.zip'),
-          'archive.zip',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() =>
-        expect(onNotification).toHaveBeenCalledWith(
-          expect.objectContaining({
-            variant: NotificationVariant.Error,
-            message: 'dialFileManager.uploadArchiveError',
-          }),
-        ),
-      );
-    });
-
-    it('resolves bucket and destinationPath relative to the destination folder', async () => {
-      mockUploadArchive.mockResolvedValue({ results: [] });
-
-      const { result } = renderHook(() =>
-        useDialFileManager({ bucket: BUCKET }),
-      );
-      await waitFor(() => expect(result.current.isLoading).toBe(false));
-
-      const file = new File(['zip'], 'archive.zip');
-      act(() => {
-        result.current.onUploadArchive(
-          file,
-          'archive.zip',
-          '/My files/reports',
-        );
-      });
-
-      await waitFor(() =>
-        expect(mockUploadArchive).toHaveBeenCalledWith(
-          file,
-          BUCKET,
-          'reports/',
-        ),
-      );
     });
   });
 });
