+++ openspec/changes/archive/2026-07-21-split-use-dial-file-manager/.openspec.yaml
